HAI-160: refactored packages and some code as well
