+++ services/hanke-service/src/integrationTest/kotlin/fi/hel/haitaton/hanke/HankeControllerITests.kt
@@ -23,7 +23,7 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 @SpringBootTest
 @AutoConfigureMockMvc
 @EnableAutoConfiguration
-class HankeControllerIntegTest(@Autowired val mockMvc: MockMvc) {
+class HankeControllerITests(@Autowired val mockMvc: MockMvc) {
 
     @Test
     fun `When hankeId not given then Bad Request`() {
