+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/HankeController.kt
@@ -1,0 +1,63 @@
+package fi.hel.haitaton.hanke
+
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.*
+import java.time.ZonedDateTime
+
+@RestController
+@RequestMapping("/hankkeet")
+class HankeController {
+
+    //TODO: get service for saving new Hanke
+
+    /**
+     * Get one hanke with hankeId.
+     *  TODO: token and user from front?
+     *  TODO: logging
+     *  TODO: validation for input
+     *  TODO: ask hankeId to be generated by an service before calling saving...or where?
+     *
+     */
+    @GetMapping
+    fun getHankeById(@RequestParam(name = "hankeId") hankeId: String?): ResponseEntity<Hanke> {
+        return if (hankeId == null) {
+            ResponseEntity.badRequest().build() //TODO, error handling
+        } else {
+            val hanke = loadHanke(hankeId)
+            if (hanke == null) {
+                ResponseEntity.notFound().build() //TODO, error handling
+            } else {
+                ResponseEntity.ok(hanke)
+            }
+        }
+    }
+
+    /**
+     * Add one hanke.
+     *  TODO: token and user from front?
+     *  TODO: logging
+     *  TODO: validation for input
+     *  TODO: ask hankeId to be generated by an service before calling saving...or where?
+     *
+     */
+    @PostMapping()
+    fun createHanke(@RequestBody hanke: Hanke?): ResponseEntity<Any> {
+        return if (hanke == null)/*TODO: can this happen?*/ {
+            ResponseEntity.badRequest().body("hanke puuttuu") //TODO, error handling
+        } else {
+            hanke.hankeId = getHankeId()
+            // TODO call service to save
+            ResponseEntity.ok(hanke)
+        }
+    }
+
+    //temporary object creator for getting started TODO: Real implementation for returning hanke from backend
+    private fun loadHanke(hankeId: String): Hanke? {
+        return Hanke(hankeId, "Mannerheimintien remontti remonttinen", ZonedDateTime.now(), ZonedDateTime.now(), "Risto", 1)
+    }
+
+    //to get new id for hankeId, TODO: this needs real implementation
+    private fun getHankeId(): String {
+        return "CREATED123"
+    }
+}
